+++ script/lib/code-sign-on-mac.js
@@ -27,7 +27,7 @@ module.exports = async function(packagedAppPath) {
   let certPath = process.env.ATOM_MAC_CODE_SIGNING_CERT_PATH;
   if (!certPath) {
     certPath = path.join(os.tmpdir(), 'mac.p12');
-    downloadFileFromGithub(
+    await downloadFileFromGithub(
       process.env.ATOM_MAC_CODE_SIGNING_CERT_DOWNLOAD_URL,
       certPath
     );